Incl dependant object deploy within Programmable Object
DROP and (re)CREATE SECURITY POLICY object that has a dependency on the Programmable Object, to ensure that changes to the FUNCTION itself can be deployed successfully.
(note that the CREATE SECURITY POLICY part of the script wasn't actually added in this commit -- this is my bad. That statement should not have been included in the "Initial commit")